requirements.txt
discord.py==1.2.5
python-dotenv
asyncio
regex
numpy==1.18.1
mechanize==0.4.5
requests==2.22.0
psycopg2==2.8.4
gspread==3.6.0
